rearrange files by add extension files
